+++ src/AutoInstaller.ts
@@ -1,8 +1,8 @@
 import * as vscode from "vscode";
 import { fixCheckSums } from "./CheckSumService";
-import { DokiSticker, DokiTheme, StickerType } from "./DokiTheme";
+import { DEFAULT_THEME_ID, DokiSticker, DokiTheme, StickerType } from "./DokiTheme";
 import DokiThemeDefinitions from "./DokiThemeDefinitions";
-import { Sticker, StickerInstallPayload } from "./extension";
+import { StickerInstallPayload } from "./extension";
 import { getHideIndex, getStickerIndex, getWallpaperIndex, hideWaterMark, InstallStatus, installStickers, installWallPaper, readCSS } from "./StickerService";
 import { getCurrentThemeAndSticker, handleInstallFailure, handleInstallMessage, showInstallNotification, showNetworkErrorMessage } from "./ThemeManager";
 
@@ -36,17 +36,23 @@ function storeFirstConfig(context: vscode.ExtensionContext) {
     saveNewVersion(context);
 
     const vscodeCSS = readCSS();
-    const { sticker } = getCurrentThemeAndSticker();
+    const { sticker, theme } = getCurrentThemeAndSticker();
     const isStickerInstalled = getStickerIndex(vscodeCSS) > -1;
     if (isStickerInstalled) {
-        saveStickerConfig(sticker, context);
+        saveStickerConfig({
+            sticker,
+            themeId: theme.id
+        }, context);
     } else {
         clearStickerConfig(context)
     }
 
     const isWallpaperInstalled = getWallpaperIndex(vscodeCSS) > -1;
     if (isWallpaperInstalled) {
-        saveWallpaperConfig(sticker, context);
+        saveWallpaperConfig({
+            sticker,
+            themeId: theme.id
+        }, context);
     } else {
         clearWallpaperConfig(context)
     }
@@ -118,7 +124,7 @@ function autoInstallAsset(
         const { theme } = getCurrentThemeAndSticker()
         const usableThemeId = themeId || theme.id;
         const def = DokiThemeDefinitions.find(theme => theme.themeDefinition.information.id === usableThemeId)
-            || DokiThemeDefinitions[0];
+            || DokiThemeDefinitions.find(theme => theme.themeDefinition.information.id === DEFAULT_THEME_ID)!;
         return assetInstaller({
             sticker: sticker.sticker,
             theme: new DokiTheme(def.themeDefinition),
@@ -129,20 +135,20 @@ function autoInstallAsset(
 }
 
 export function saveStickerConfig(
-    sticker: DokiSticker,
+    restoreConfig: RestoreConfig,
     context: vscode.ExtensionContext,
 ) {
     context.globalState.update(
-        stickerInstallKey, createAssetRestoreConfig(sticker)
+        stickerInstallKey, createAssetRestoreConfig(restoreConfig)
     );
 }
 
 export function saveWallpaperConfig(
-    sticker: DokiSticker,
+    restoreConfig: RestoreConfig,
     context: vscode.ExtensionContext,
 ) {
     context.globalState.update(
-        wallpaperInstallKey, createAssetRestoreConfig(sticker)
+        wallpaperInstallKey, createAssetRestoreConfig(restoreConfig)
     )
 }
 
@@ -156,7 +162,10 @@ export function saveHiddenWatermarkConfig(context: vscode.ExtensionContext) {
         type: StickerType.DEFAULT,
     }
     context.globalState.update(
-        watermarkKey, createAssetRestoreConfig(bestSticker),
+        watermarkKey, createAssetRestoreConfig({
+            sticker: bestSticker,
+            themeId: DEFAULT_THEME_ID,
+        }),
     )
 }
 
@@ -178,17 +187,13 @@ export function clearAssetConfig(
     clearWatermarkConfig(context);
 }
 
-type RestoreConfig = {
+export type RestoreConfig = {
     sticker: DokiSticker;
     themeId: string;
 }
 
-function createAssetRestoreConfig(sticker: DokiSticker): string {
-    return JSON.stringify(
-        {
-            sticker
-        } as RestoreConfig
-    )
+function createAssetRestoreConfig(stickerInstallPayload: RestoreConfig): string {
+    return JSON.stringify(stickerInstallPayload)
 }
 
 function wasAssetInstalled(