+++ CHANGELOG.md
@@ -18,6 +18,8 @@ _Zero Two's Not Just A Cutie. ;)_
 
 - Desaturated Itsuki's theme a bit.
 - It was brought to my attention that I cannot spell "Rimuru".
+- Zero Two's Obsidian theme is the default theme now.
+
 
 # 84.2-1.3.0 [Enhancements n Stuff]